+++ constants.go
@@ -1,0 +1,12 @@
+// Copyright 2024 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+package terramate
+
+const (
+	// DefaultFilename is the name of the default Terramate configuration file.
+	DefaultFilename = "terramate.tm.hcl"
+
+	// SkipFilename is the name of Terramate skip file.
+	SkipFilename = ".tmskip"
+)